feat: replace min-words with min-chars
Prior to this change, `findlike` counted the number of words in each document by parsing the words with regex. This step was done inside the `Corpus` object.

However, tokenization also took place afterwards in the `Processor` object.

Therefore, to reduce the unnecessary overhead, I decided to replace word counting with character counting.